+++ joblib/_dask.py
@@ -8,8 +8,8 @@ import time
 from uuid import uuid4
 import weakref
 
-from .parallel import AutoBatchingMixin, ParallelBackendBase, BatchedCalls
 from .parallel import parallel_backend
+from .parallel import AutoBatchingMixin, ParallelBackendBase
 
 try:
     import dask
@@ -19,7 +19,7 @@ except ImportError:
     distributed = None
 
 if dask is not None and distributed is not None:
-    from dask.utils import funcname, itemgetter
+    from dask.utils import funcname
     from dask.sizeof import sizeof
     from dask.distributed import (
         Client,
@@ -27,10 +27,8 @@ if dask is not None and distributed is not None:
         get_client,
         secede,
         rejoin,
-        get_worker
     )
     from distributed.utils import thread_state
-
 
     try:
         # asyncio.TimeoutError, Python3-only error thrown by recent versions of
@@ -126,7 +124,7 @@ class Batch:
         with parallel_backend('dask'):
             for func, args, kwargs in tasks:
                 results.append(func(*args, **kwargs))
-        return results
+            return results
 
     def __repr__(self):
         descr = f"batch_of_{self._funcname}_{self._num_tasks}_calls"