+++ joblib/logger.py
@@ -76,6 +76,9 @@ class Logger(object):
     def warn(self, msg):
         logging.warning("[%s]: %s" % (self, msg))
 
+    def info(self, msg):
+        logging.info("[%s]: %s" % (self, msg))
+
     def debug(self, msg):
         # XXX: This conflicts with the debug flag used in children class
         logging.debug("[%s]: %s" % (self, msg))
@@ -109,19 +112,19 @@ class PrintTime(object):
                     try:
                         shutil.move(logfile + '.%i' % i,
                                     logfile + '.%i' % (i + 1))
-                    except:
+                    except:  # noqa: E722
                         "No reason failing here"
                 # Use a copy rather than a move, so that a process
                 # monitoring this file does not get lost.
                 try:
                     shutil.copy(logfile, logfile + '.1')
-                except:
+                except:  # noqa: E722
                     "No reason failing here"
             try:
                 with open(logfile, 'w') as logfile:
                     logfile.write('\nLogging joblib python script\n')
                     logfile.write('\n---%s---\n' % time.ctime(self.last_time))
-            except:
+            except:  # noqa: E722
                 """ Multiprocessing writing to files can create race
                     conditions. Rather fail silently than crash the
                     computation.
@@ -146,7 +149,7 @@ class PrintTime(object):
             try:
                 with open(self.logfile, 'a') as f:
                     print(full_msg, file=f)
-            except:
+            except:  # noqa: E722
                 """ Multiprocessing writing to files can create race
                     conditions. Rather fail silently than crash the
                     calculation.