+++ joblib/memory.py
@@ -10,7 +10,9 @@ is called with the same input arguments.
 
 
 from __future__ import with_statement
+import logging
 import os
+from textwrap import dedent
 import time
 import pathlib
 import pydoc
@@ -440,7 +442,7 @@ class MemorizedFunc(Logger):
         self.timestamp = timestamp
         try:
             functools.update_wrapper(self, func)
-        except:
+        except:  # noqa: E722
             " Objects like ufunc don't like that "
         if inspect.isfunction(func):
             doc = pydoc.TextDoc().document(func)
@@ -490,6 +492,26 @@ class MemorizedFunc(Logger):
 
         # Whether or not the memorized function must be called
         must_call = False
+
+        if self._verbose >= 20:
+            logging.basicConfig(level=logging.INFO)
+            _, name = get_func_name(self.func)
+            location = self.store_backend.get_cached_func_info([func_id])[
+                'location']
+            _, signature = format_signature(self.func, *args, **kwargs)
+
+            self.info(
+                dedent(
+                    f"""
+                        Querying {name} with signature
+                        {signature}.
+
+                        (argument hash {args_id})
+
+                        The store location is {location}.
+                        """
+                )
+            )
 
         # FIXME: The statements below should be try/excepted
         # Compare the function code with the previous to see if the
@@ -843,16 +865,12 @@ class MemorizedFunc(Logger):
             # for which repr() always output a short representation, but can
             # be with complex dictionaries. Fixing the problem should be a
             # matter of replacing repr() above by something smarter.
-            warnings.warn("Persisting input arguments took %.2fs to run.\n"
+            warnings.warn("Persisting input arguments took %.2fs to run."
                           "If this happens often in your code, it can cause "
-                          "performance problems \n"
-                          "(results will be correct in all cases). \n"
+                          "performance problems "
+                          "(results will be correct in all cases). "
                           "The reason for this is probably some large input "
-                          "arguments for a wrapped\n"
-                          " function (e.g. large strings).\n"
-                          "THIS IS A JOBLIB ISSUE. If you can, kindly provide "
-                          "the joblib's team with an\n"
-                          " example so that they can fix the problem."
+                          "arguments for a wrapped function."
                           % this_duration, stacklevel=5)
         return metadata