+++ joblib/test/test_memory.py
@@ -8,6 +8,7 @@ Test the memory module.
 
 import functools
 import gc
+import logging
 import shutil
 import os
 import os.path
@@ -33,7 +34,6 @@ from joblib.test.common import with_numpy, np
 from joblib.test.common import with_multiprocessing
 from joblib.testing import parametrize, raises, warns
 from joblib.hashing import hash
-
 
 
 ###############################################################################
@@ -91,9 +91,9 @@ def test_memory_integration(tmpdir):
     # thus it serves as a test to see that both are identified
     # as different.
 
-    def f(l):
+    def f(arg):
         accumulator.append(1)
-        return l
+        return arg
 
     check_identity_lazy(f, accumulator, tmpdir.strpath)
 
@@ -229,9 +229,9 @@ def test_no_memory():
     """ Test memory with location=None: no memoize """
     accumulator = list()
 
-    def ff(l):
+    def ff(arg):
         accumulator.append(1)
-        return l
+        return arg
 
     memory = Memory(location=None, verbose=0)
     gg = memory.cache(ff)
@@ -245,16 +245,16 @@ def test_memory_kwarg(tmpdir):
     " Test memory with a function with keyword arguments."
     accumulator = list()
 
-    def g(l=None, m=1):
+    def g(arg1=None, arg2=1):
         accumulator.append(1)
-        return l
+        return arg1
 
     check_identity_lazy(g, accumulator, tmpdir.strpath)
 
     memory = Memory(location=tmpdir.strpath, verbose=0)
     g = memory.cache(g)
     # Smoke test with an explicit keyword argument:
-    assert g(l=30, m=2) == 30
+    assert g(arg1=30, arg2=2) == 30
 
 
 def test_memory_lambda(tmpdir):
@@ -267,9 +267,7 @@ def test_memory_lambda(tmpdir):
         accumulator.append(1)
         return x
 
-    l = lambda x: helper(x)
-
-    check_identity_lazy(l, accumulator, tmpdir.strpath)
+    check_identity_lazy(lambda x: helper(x), accumulator, tmpdir.strpath)
 
 
 def test_memory_name_collision(tmpdir):
@@ -303,10 +301,8 @@ def test_memory_name_collision(tmpdir):
 def test_memory_warning_lambda_collisions(tmpdir):
     # Check that multiple use of lambda will raise collisions
     memory = Memory(location=tmpdir.strpath, verbose=0)
-    a = lambda x: x
-    a = memory.cache(a)
-    b = lambda x: x + 1
-    b = memory.cache(b)
+    a = memory.cache(lambda x: x)
+    b = memory.cache(lambda x: x + 1)
 
     with warns(JobLibCollisionWarning) as warninfo:
         assert a(0) == 0
@@ -392,9 +388,9 @@ def test_memory_numpy(tmpdir, mmap_mode):
     " Test memory with a function with numpy arrays."
     accumulator = list()
 
-    def n(l=None):
+    def n(arg=None):
         accumulator.append(1)
-        return l
+        return arg
 
     memory = Memory(location=tmpdir.strpath, mmap_mode=mmap_mode,
                     verbose=0)
@@ -1314,3 +1310,28 @@ def test_memory_pickle_dump_load(tmpdir, memory_kwargs):
     compare(memorized_result, memorized_result_reloaded,
             ignored_attrs=set(['store_backend', 'timestamp', '_func_code_id']))
     assert hash(memorized_result) == hash(memorized_result_reloaded)
+
+
+def test_info_log(tmpdir, caplog):
+    caplog.set_level(logging.INFO)
+    x = 3
+
+    memory = Memory(location=tmpdir.strpath, verbose=20)
+
+    @memory.cache
+    def f(x):
+        return x ** 2
+
+    _ = f(x)
+    assert "Querying" in caplog.text
+    caplog.clear()
+
+    memory = Memory(location=tmpdir.strpath, verbose=0)
+
+    @memory.cache
+    def f(x):
+        return x ** 2
+
+    _ = f(x)
+    assert "Querying" not in caplog.text
+    caplog.clear()